Move FileDescEnvironment::report_error to its own environment
* Mainstream shells will report errors by prefixing their own name
before the error. To emulate the same in `Env`, we would be forced to
only implement FileDescEnvironment if we also implement
ArgumentsEnvironment (so that we can grab the shell's name).
* It does not appear appropriate to couple the FileDescEnvironment impl
on `Env` with ArgumentsEnvironment, and since specialization is not yet
stable, we don't have much luck there
* By splitting it out into its own environment, `Env` can implement it
only if it is both ArgumentsEnvironment + FileDescEnvironment, without
coupling the two together